moved counter file to components folder
